fixes for issues in capi-garbage-cleanup.sh

+++ kustomize-units/cluster-garbage-collector/capi-garbage-cleanup.sh
@@ -3,13 +3,15 @@
 set -ue
 set -o pipefail
 
+# the '.*' below is here to ensure a match in `kubectl api-resources | grep ...`
+# we use fully-qualified kinds to avoid unfortunate collisions with Rancher CRDs
 TEMPLATE_TYPES_CR="
-KubeadmConfigTemplates
-RKE2ConfigTemplates
-DockerMachineTemplates
-VSphereMachineTemplates.*cluster.x-k8s.io
-OpenStackMachineTemplates
-Metal3MachineTemplates
+KubeadmConfigTemplate.*bootstrap.cluster.x-k8s.io
+RKE2ConfigTemplate.*bootstrap.cluster.x-k8s.io
+DockerMachineTemplate.*infrastructure.cluster.x-k8s.io
+VSphereMachineTemplate.*infrastructure.cluster.x-k8s.io
+OpenStackMachineTemplate.*infrastructure.cluster.x-k8s.io
+Metal3MachineTemplate.*infrastructure.cluster.x-k8s.io
 "
 
 for TEMPLATE_CR in ${TEMPLATE_TYPES_CR[@]}; do
@@ -20,21 +22,25 @@ for TEMPLATE_CR in ${TEMPLATE_TYPES_CR[@]}; do
         # (see VSphereMachineTemplates.*cluster.x-k8s.io above)
         TEMPLATE_RESOURCE=${TEMPLATE_CR/\*/}
 
+        # in some places we need to use the short kind name
+        TEMPLATE_RESOURCE_SHORT=${TEMPLATE_CR/.*/}
+
         # Only for (Docker|VSphere|OpenStack|Metal3)MachineTemplates resources
         if [[ $TEMPLATE_RESOURCE =~ .*MachineTemplate ]]; then
 
             # Extract cloned template resource type from template type (eg. OpenStackMachine out of OpenStackMachineTemplate)
             CLONED_TEMPLATE_RESOURCE="${TEMPLATE_RESOURCE/Template}"
+            CLONED_TEMPLATE_RESOURCE_SHORT=${CLONED_TEMPLATE_RESOURCE/.*/}
 
-            echo -e "\n Inspecting all $TEMPLATE_RESOURCE resource instances/names for
+            echo -e "\n Inspecting all $TEMPLATE_RESOURCE_SHORT resource instances/names for
 - usage as (RKE2ControlPlane.spec|KubeadmControlPlane.spec.machineTemplate).infrastructureRef.name
 - usage as (MachineDeployment|MachineSet).spec.template.spec.infrastructureRef.name
-- presence of $CLONED_TEMPLATE_RESOURCE resources having the annotation cluster.x-k8s.io/cloned-from-name=<$TEMPLATE_RESOURCE resource instance>"
+- presence of $CLONED_TEMPLATE_RESOURCE_SHORT resources having the annotation cluster.x-k8s.io/cloned-from-name=<$TEMPLATE_RESOURCE_SHORT resource instance>"
 
         # Only for (Kubeadm|RKE2)ConfigTemplate resources
         elif [[ $TEMPLATE_RESOURCE =~ .*ConfigTemplate ]]; then
 
-            echo -e "\n Inspecting all $TEMPLATE_RESOURCE resource instances/names for
+            echo -e "\n Inspecting all $TEMPLATE_RESOURCE_SHORT resource instances/names for
 - usage as (MachineDeployment|MachineSet).spec.template.spec.bootstrap.configRef.name"
 
         fi
@@ -47,7 +53,7 @@ for TEMPLATE_CR in ${TEMPLATE_TYPES_CR[@]}; do
             TEMPLATE_RESOURCE_INSTANCE_LIST=$(kubectl -n "$TARGET_NAMESPACE" get "$TEMPLATE_RESOURCE" -o json | jq -r --argjson timestamp 300 '.items[] | select (.metadata.creationTimestamp | sub("\\..*";"Z") | sub("\\s";"T") | fromdate < now - $timestamp).metadata.name')
 
             if [ -z "$TEMPLATE_RESOURCE_INSTANCE_LIST" ]; then
-                echo -e "\n\t All $TEMPLATE_RESOURCE inside namespace $TARGET_NAMESPACE are newer than 5 minutes \n"
+                echo -e "\n\t All $TEMPLATE_RESOURCE_SHORT inside namespace $TARGET_NAMESPACE are newer than 5 minutes \n"
                 kubectl -n "$TARGET_NAMESPACE" get "$TEMPLATE_RESOURCE" -o=custom-columns=KIND:.kind,NAME:.metadata.name,NAMESPACE:.metadata.namespace,AGE:.metadata.creationTimestamp
                 continue
             fi
@@ -77,8 +83,8 @@ for TEMPLATE_CR in ${TEMPLATE_TYPES_CR[@]}; do
                     kubectl -n "$TARGET_NAMESPACE" get MachineSet -o=custom-columns=KIND:.kind,NAME:.metadata.name,INFRASTRUCTURE_REF_NAME:.spec.template.spec.infrastructureRef.name,INFRASTRUCTURE_REF_KIND:.spec.template.spec.infrastructureRef.kind --no-headers >> /tmp/template_resources_consumers.txt
 
                     while read INFRASTRUCTURE_CONSUMER_KIND INFRASTRUCTURE_CONSUMER_INSTANCE INFRASTRUCTURE_REF_NAME INFRASTRUCTURE_REF_KIND; do
-                    if [[ $INFRASTRUCTURE_REF_NAME == $TEMPLATE_RESOURCE_INSTANCE && $INFRASTRUCTURE_REF_KIND == $TEMPLATE_RESOURCE ]]; then
-                        echo -e "\n\t Found $TEMPLATE_RESOURCE/$TEMPLATE_RESOURCE_INSTANCE being referenced by $INFRASTRUCTURE_CONSUMER_KIND/$INFRASTRUCTURE_CONSUMER_INSTANCE."
+                    if [[ $INFRASTRUCTURE_REF_NAME == $TEMPLATE_RESOURCE_INSTANCE && $INFRASTRUCTURE_REF_KIND == $TEMPLATE_RESOURCE_SHORT ]]; then
+                        echo -e "\n\t Found $TEMPLATE_RESOURCE_SHORT/$TEMPLATE_RESOURCE_INSTANCE being referenced by $INFRASTRUCTURE_CONSUMER_KIND/$INFRASTRUCTURE_CONSUMER_INSTANCE."
                         could_be_removed="false"
                     fi
                     done < /tmp/template_resources_consumers.txt
@@ -88,9 +94,9 @@ for TEMPLATE_CR in ${TEMPLATE_TYPES_CR[@]}; do
 
                     # Don't consider template resource instance for deletion if cloned resources still exist for it
                     if [[ $EXISTING_CLONED_RESOURCES -gt 0 ]]; then
-                        echo -e "\n\t The following $CLONED_TEMPLATE_RESOURCE resources are annotated with cluster.x-k8s.io/cloned-from-name=$TEMPLATE_RESOURCE_INSTANCE"
+                        echo -e "\n\t The following $CLONED_TEMPLATE_RESOURCE_SHORT resources are annotated with cluster.x-k8s.io/cloned-from-name=$TEMPLATE_RESOURCE_INSTANCE:"
                         kubectl -n "$TARGET_NAMESPACE" get $CLONED_TEMPLATE_RESOURCE -o jsonpath="{ $.items[?(@.metadata.annotations.cluster\.x-k8s\.io\/cloned-from-name == '$TEMPLATE_RESOURCE_INSTANCE')].metadata.name}" | sed 's/^/\'$'\t/'
-                        echo -e "\n\t Thus $TEMPLATE_RESOURCE/$TEMPLATE_RESOURCE_INSTANCE is not considered for garbage collection."
+                        echo -e "\n\t Thus $TEMPLATE_RESOURCE_SHORT/$TEMPLATE_RESOURCE_INSTANCE is not considered for garbage collection."
                         could_be_removed="false"
                     fi
                 fi
@@ -105,16 +111,48 @@ for TEMPLATE_CR in ${TEMPLATE_TYPES_CR[@]}; do
                     kubectl -n "$TARGET_NAMESPACE" get MachineDeployment -o=custom-columns=KIND:.kind,NAME:.metadata.name,BOOTSTRAP_CONFIG_REF_NAME:.spec.template.spec.bootstrap.configRef.name,BOOTSTRAP_CONFIG_REF_KIND:.spec.template.spec.bootstrap.configRef.kind --no-headers >> /tmp/template_resources_consumers.txt
 
                     while read BOOTSTRAP_CONSUMER_KIND BOOTSTRAP_CONSUMER_INSTANCE BOOTSTRAP_CONFIG_REF_NAME BOOTSTRAP_CONFIG_REF_KIND; do
-                    if [[ $BOOTSTRAP_CONFIG_REF_NAME == $TEMPLATE_RESOURCE_INSTANCE && $BOOTSTRAP_CONFIG_REF_KIND == $TEMPLATE_RESOURCE ]]; then
-                        echo -e "\n\t Found $TEMPLATE_RESOURCE/$TEMPLATE_RESOURCE_INSTANCE being referenced by $BOOTSTRAP_CONSUMER_KIND/$BOOTSTRAP_CONSUMER_INSTANCE."
-                        could_be_removed="false"
-                    fi
+                        if [[ $BOOTSTRAP_CONFIG_REF_NAME == $TEMPLATE_RESOURCE_INSTANCE && $BOOTSTRAP_CONFIG_REF_KIND == $TEMPLATE_RESOURCE_SHORT ]]; then
+                            echo -e "\n\t Found $TEMPLATE_RESOURCE_SHORT/$TEMPLATE_RESOURCE_INSTANCE being referenced by $BOOTSTRAP_CONSUMER_KIND/$BOOTSTRAP_CONSUMER_INSTANCE."
+                            could_be_removed="false"
+                        fi
                     done < /tmp/template_resources_consumers.txt
                 fi
 
                 if [[ "$could_be_removed" == "true" ]]; then
-                    echo -e "\n\t Deleting $TEMPLATE_RESOURCE/$TEMPLATE_RESOURCE_INSTANCE"
-                    kubectl -n "$TARGET_NAMESPACE" delete "$TEMPLATE_RESOURCE/$TEMPLATE_RESOURCE_INSTANCE" | sed 's/^/\'$'\t/'
+                    echo -e "\n\t Preparing deletion of $TEMPLATE_RESOURCE_SHORT/$TEMPLATE_RESOURCE_INSTANCE"
+                    owner_ref_count=$(kubectl -n "$TARGET_NAMESPACE" get "$TEMPLATE_RESOURCE/$TEMPLATE_RESOURCE_INSTANCE" -o yaml | yq '.metadata.ownerReferences | length')
+                    if [[ $owner_ref_count != "0" ]]; then
+                        echo -e "\n\t **NOT** deleting $TEMPLATE_RESOURCE_SHORT/$TEMPLATE_RESOURCE_INSTANCE, because it still has ownerReferences (anomaly in this tool ?)"
+                        kubectl -n "$TARGET_NAMESPACE" get "$TEMPLATE_RESOURCE/$TEMPLATE_RESOURCE_INSTANCE" -o yaml | yq '.metadata.ownerReferences'
+                    else
+                        TEMPLATE_RESOURCE_API_VERSION=$(kubectl -n "$TARGET_NAMESPACE" get "$TEMPLATE_RESOURCE/$TEMPLATE_RESOURCE_INSTANCE" -o yaml | yq .apiVersion)
+
+                        echo -e "\n\t Deleting $TEMPLATE_RESOURCE_SHORT/$TEMPLATE_RESOURCE_INSTANCE"
+                        kubectl -n "$TARGET_NAMESPACE" delete "$TEMPLATE_RESOURCE/$TEMPLATE_RESOURCE_INSTANCE" | sed 's/^/\'$'\t/'
+
+                        echo -e "\n\t Creating an event for this deletion"
+                        kubectl apply -f - <<EOF
+apiVersion: v1
+kind: Event
+metadata:
+  name: capi-garbage-cleanup.delete.$(printf "%0000x%0000x" $RANDOM $RANDOM)
+  namespace: $TARGET_NAMESPACE
+type: Normal
+reportingComponent: capi-garbage-cleanup
+message: capi-garbage-cleanup deleted $TEMPLATE_RESOURCE/$TEMPLATE_RESOURCE_INSTANCE
+reason: capi-garbage-cleanup determined that this object isn't referred to anymore
+firstTimestamp: $(date -Iseconds -u | sed 's/+00:00/Z/')
+involvedObject:*
+  apiVersion: $TEMPLATE_RESOURCE_API_VERSION
+  kind: $TEMPLATE_RESOURCE
+  name: $TEMPLATE_RESOURCE_INSTANCE
+  namespace: $TARGET_NAMESPACE
+action: delete
+source:
+  component: capi-garbage-cleanup
+  host: $HOSTNAME
+EOF
+                    fi
                 fi
             done
         done
